refactor: LessonRoundItemResponse 가격관련 필드 Int -> Long으로 수정

+++ src/main/kotlin/team/mozu/dsm/adapter/in/lesson/dto/response/LessonRoundItemResponse.kt
@@ -4,7 +4,7 @@ data class LessonRoundItemResponse(
     val itemId: Int,
     val itemName: String,
     val itemLogo: String?,
-    val nowMoney: Int,
-    val profitMoney: Int,
+    val nowMoney: Long,
+    val profitMoney: Long,
     val profitNum: Double
 )
